Протоколы отдельно от имплеминтаций
