+++ MovieQuiz/MovieQuizModel/GameState.swift
@@ -1,5 +1,5 @@
 enum GameState {
-    case nextRiddle(riddle: MovieRiddleProtocol, riddleNum: Int, riddlesCount: Int)
+    case nextRiddle(riddle: MovieRiddle, riddleNum: Int, riddlesCount: Int)
     case positiveAnswer
     case negativeAnswer
     case gameEnded(correctAnswers: Int, riddlesCount: Int)
